+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - added more information to the debug tab in the settings
 - the todo tab in the settings is now disabled if there are no ownCloud settings
   (for [Issue #208](https://github.com/pbek/QOwnNotes/issues/208))
+- the app instance settings are now hidden on OS X since we don't need them 
+  there (for [Issue #208](https://github.com/pbek/QOwnNotes/issues/208))
 
 ## 16.06.6
 - you now can use a real **dark mode** on every operating system